Move RoomPos to its own file and rename Positioning
